refactor(file-structure): split project graphql inputs
