+++ apps/project/graphql/inputs/inputs.py
@@ -1,0 +1,83 @@
+import strawberry
+import strawberry_django
+from strawberry.file_uploads import Upload
+
+from apps.common.graphql.inputs import (
+    UserResourceCreateInputMixin,
+    UserResourceTopLevelUpdateInputMixin,
+)
+from apps.project.models import Organization, Project, ProjectAsset
+
+from .project_types.compare import CompareProjectPropertyInput
+from .project_types.completeness import CompletenessProjectPropertyInput
+from .project_types.find import FindProjectPropertyInput
+from .project_types.validate import ValidateProjectPropertyInput
+
+
+# Organization
+@strawberry_django.input(Organization)
+class OrganizationCreateInput(UserResourceCreateInputMixin):
+    name: strawberry.auto
+
+
+@strawberry_django.input(Organization)
+class OrganizationUpdateInput(UserResourceTopLevelUpdateInputMixin):
+    name: strawberry.auto
+
+
+# Project Properties
+@strawberry.input(one_of=True)
+class ProjectTypeSpecificInput:
+    compare: CompareProjectPropertyInput | None = strawberry.UNSET
+    find: FindProjectPropertyInput | None = strawberry.UNSET
+    completeness: CompletenessProjectPropertyInput | None = strawberry.UNSET
+    validate: ValidateProjectPropertyInput | None = strawberry.UNSET
+
+
+# NOTE: Make sure this matches with the serializers ../serializers.py
+@strawberry_django.input(Project)
+class ProjectCreateInput(UserResourceCreateInputMixin):
+    project_type: strawberry.auto
+    requesting_organization: strawberry.ID
+    name: strawberry.auto
+    look_for: strawberry.auto
+    additional_info_url: strawberry.auto
+    description: strawberry.auto
+
+
+# NOTE: Make sure this matches with the serializers ../serializers.py
+@strawberry_django.partial(Project)
+class ProjectUpdateInput(UserResourceTopLevelUpdateInputMixin):
+    name: strawberry.auto
+    look_for: strawberry.auto
+    additional_info_url: strawberry.auto
+    description: strawberry.auto
+    verification_number: strawberry.auto
+    group_size: strawberry.auto
+    max_tasks_per_user: strawberry.auto
+    status: strawberry.auto
+    tutorial: strawberry.ID | None = strawberry.UNSET
+    requesting_organization: strawberry.ID | None = strawberry.UNSET
+    image: strawberry.ID | None = strawberry.UNSET
+    project_type_specifics: ProjectTypeSpecificInput | None = strawberry.UNSET
+
+
+# NOTE: Make sure this matches with the serializers ../serializers.py
+@strawberry_django.partial(Project)
+class ProcessedProjectUpdateInput(UserResourceTopLevelUpdateInputMixin):
+    name: strawberry.auto
+    look_for: strawberry.auto
+    additional_info_url: strawberry.auto
+    description: strawberry.auto
+    status: strawberry.auto
+    tutorial: strawberry.ID | None = strawberry.UNSET
+    requesting_organization: strawberry.ID | None = strawberry.UNSET
+    image: strawberry.ID | None = strawberry.UNSET
+
+
+# NOTE: Make sure this matches with the serializers ../serializers.py
+@strawberry_django.input(ProjectAsset)
+class ProjectAssetCreateInput(UserResourceCreateInputMixin):
+    mimetype: strawberry.auto
+    file: Upload
+    project: strawberry.ID
