+++ apps/project/graphql/queries.py
@@ -8,7 +8,7 @@ from utils.geo.vector_tile_server.config import VectorConfig, VectorTileServerNa
 
 from .filters import OrganizationFilter, ProjectAssetFilter, ProjectFilter
 from .orders import OrganizationOrder, ProjectAssetOrder, ProjectOrder
-from .types import (
+from .types.types import (
     OrganizationType,
     ProjectAssetType,
     ProjectType,
